move work_queue specific code to work_queue_tools.py
